Refactor scene graph to remove render logic from components and put it in the renderer itself.
